feature: #279 Confirmed basic concepts work by adding integration tests
* Split up modules so that it's easier to setup the integration test without running the app
* Removed the health endpoint from the api spec since it's added automatically
* Removed the result group tag because it didn't seem to work (at least in the integration test)
* Added fixes for getting environment variables into the file as well as for the routes to be setup correctly
* Fixed reference to extension since they no longer come back as a RawMessage
* Changed the sequence number to int64
* Fixed issue where event handlers were not being triggered
* Added api config to the middleware params
* Updated the command dispatcher / handler signature so that the response is value based (not a pointer)

+++ v2/rest/event.go
@@ -46,7 +46,7 @@ func (e *Event) GetType() string {
 	panic("implement me")
 }
 
-func (e *Event) GetSequenceNo() int {
+func (e *Event) GetSequenceNo() int64 {
 	//TODO implement me
 	panic("implement me")
 }
